# openapi/openapi.yaml
openapi: 3.0.0
info:
  title: API
  version: 1.0.0
paths:
  /tasks:
    get:
      summary: Получить список всех задач
      tags:
        - tasks
      description: Возвращает список всех задач из базы данных.
      responses:
        '200':
          description: Успешный ответ. Возвращает массив задач.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
        '500':
          $ref: '#/components/responses/InternalServerError'  # Используем переиспользуемый ответ

    post:
      summary: Создать новую задачу
      tags:
        - tasks
      description: Создает новую задачу в базе данных.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                task:
                  type: string
                  description: Описание задачи
                is_done:
                  type: boolean
                  description:  c
              required:
                - task
      responses:
        '201':
          description: Задача успешно создана. Возвращает созданную задачу.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          $ref: '#/components/responses/BadRequestError'  # Используем переиспользуемый ответ
        '500':
          $ref: '#/components/responses/InternalServerError'  # Используем переиспользуемый ответ
  /tasks/{id}:
    put: # or patch
      summary: Обновить существующую задачу
      tags:
        - tasks
      description: Обновляет задачу с указанным ID.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
          description: ID задачи для обновления
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Task'
      responses:
        '200':
          description: Задача успешно обновлена. Возвращает обновленную задачу.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '404':
          description: Задача не найдена.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      summary: Удалить задачу
      tags:
        - tasks
      description: Удаляет задачу с указанным ID.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
          description: ID задачи для удаления
      responses:
        '204':
          description: Задача успешно удалена (No Content).
        '404':
          description: Задача не найдена.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/InternalServerError'
components:
  schemas:
    Task:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Уникальный идентификатор задачи
          readOnly: true  # ID генерируется сервером
        task:
          type: string
          description: Описание задачи
        is_done:
          type: boolean
          description: Статус выполнения задачи (true - выполнена, false - не выполнена)
      required:
        - task

    Error:
      type: object
      properties:
        error:
          type: string
          description: Сообщение об ошибке

  responses:
    BadRequestError:
      description: Некорректный запрос
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalServerError:
      description: Ошибка сервера
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'